+++ drivers/KS-Delta/ks_lib.cpp
