+++ drivers/Lorenz-Delta/Makefile
@@ -26,17 +26,14 @@
 # Import machine specific compilers, options, flags, etc.. 
 ##################################################################
 
-BRAID_DIR=../braid
-include ../makefile.inc
+BRAID_DIR=../../braid
+include ../../makefile.inc
 
 
 ##################################################################
-# Build exmaples 
+# Build
 ##################################################################
 
-HYPRE_DIR = ../../../hypre/src/hypre
-HYPRE_FLAGS = -I$(HYPRE_DIR)/include
-HYPRE_LIB = -L$(HYPRE_DIR)/lib -lHYPRE
 SUITE_SPARSE = /usr/include/suitesparse/
 
 MFEM_INTERFACE = ./braid_mfem.hpp ./braid_mfem_block.hpp ../braid/braid.hpp
@@ -48,12 +45,6 @@ MFEM_LIB_FILE = mfem_is_not_built
 BRAID_FLAGS = -I$(BRAID_DIR) -Wno-deprecated-declarations
 BRAID_LIB_FILE = $(BRAID_DIR)/libbraid.a
 
-C_EXAMPLES = drive-burgers-1D drive-diffusion-2D drive-solve-adjoint-with-xbraid
-# Note: .cpp examples will be linked with mfem
-CXX_EXAMPLES = drive-diffusion-1D-moving-mesh drive-diffusion-1D-moving-mesh-serial \
-					drive-diffusion-ben drive-diffusion drive-lin-elasticity \
-					drive-pLaplacian drive-adv-diff-DG drive-nonlin-elasticity
-
 # clean all possible
 ALL_EXAMPLES := $(patsubst %.cpp,%,$(wildcard *.c *.cpp))
 ALL_EXAMPLES := $(patsubst %.c,%,$(ALL_EXAMPLES))
@@ -63,24 +54,15 @@ ALL_EXAMPLES := $(patsubst %.c,%,$(ALL_EXAMPLES))
 .SUFFIXES:
 .SUFFIXES: .c .cpp
 
-# put this rule first so it becomes the default
-all: $(C_NOHYPRE) $(C_EXAMPLES) $(CXX_EXAMPLES) $(CXX_MFEM_EXAMPLES)
-	
-# Rule for compiling .c files
-%: %.c $(BRAID_LIB_FILE)
+drive-lorenz-Delta: drive-lorenz-Delta.cpp $(BRAID_LIB_FILE)
 	@echo "Building" $@ "..."
-	$(MPICC) $(CFLAGS) $(BRAID_FLAGS) $(HYPRE_FLAGS) $< -o $@\
- $(BRAID_LIB_FILE) $(HYPRE_LIB) $(LFLAGS)
+	$(MPICXX) $(CXXFLAGS) $(BRAID_FLAGS) -Idrive-lorenz-Delta-lib -IEigen \
+ $< -o $@ $(BRAID_LIB_FILE) lorenz_lib.cpp $(LFLAGS)
 
-# Rule for compiling .cpp files; links with mfem
-%: %.cpp $(BRAID_LIB_FILE) $(MFEM_LIB_FILE) $(MFEM_CONFIG_MK) $(MFEM_INTERFACE)
+drive-lorenz-LRDelta: drive-lorenz-LRDelta.cpp $(BRAID_LIB_FILE)
 	@echo "Building" $@ "..."
-	$(MPICXX) $(CXXFLAGS) $(BRAID_FLAGS) $(MFEM_FLAGS) \
- $< -o $@ $(MFEM_LIBS) $(BRAID_LIB_FILE) $(LFLAGS)
-
-# Generate an error message if the MFEM library is not built and exit
-$(MFEM_LIB_FILE):
-	$(error The MFEM library is not built)
+	$(MPICXX) $(CXXFLAGS) $(BRAID_FLAGS) -Idrive-lorenz-Delta-lib -IEigen \
+ $< -o $@ $(BRAID_LIB_FILE) lorenz_lib.cpp $(LFLAGS)
 
 clean: cleanout
 	rm -f *.o $(ALL_EXAMPLES) *ror_norm* *_err_* *_sol_*
